Make modifier grabs configurable
The original code only grabs Control, Alt, Control+Shift and Alt+Shift.
While he lack of versatility is slightly upsetting, the real problem is
that NumLock is considered a modifier so turning it on breaks the
grabbing, confusing some people.

This, for example, is a likely victim:
http://askubuntu.com/questions/143573

This commit allows the user to specify the modifiers they want to grab
when the program is invoked.

This fix only applies to the unix binary!

I also moved the in-code header documentation to a README since it was
getting a little long.