Listen to keyboard layout changes.
The Unix binary now automatically realizes whether the current
layout is Dvorak and grabs/ungrabs keys as appropriate. This
prevents the binary from shuffling the shortcuts all over the place
when the user reverts to Qwerty, for example.

This is part of kentonv/dvorak-qwerty#2. (It's missing the
Windows implementation)

The code also validates argvs better. It now complains if the user
sends unknown modifiers as program arguments. These bogus strings
used to be simply ignored.